move table utils
