+++ assets/src/components/backups/cluster/backups/Backups.tsx
@@ -9,7 +9,7 @@ import { ComponentProps, useMemo } from 'react'
 import isEmpty from 'lodash/isEmpty'
 import { useParams } from 'react-router-dom'
 
-import { useFetchPaginatedData } from 'components/cd/utils/useFetchPaginatedData'
+import { useFetchPaginatedData } from 'components/utils/table/useFetchPaginatedData'
 
 import {
   useClusterBackupsQuery,
