+++ assets/src/components/cd/clusters/Clusters.tsx
@@ -50,7 +50,7 @@ import {
 
 import { TagsFilter } from '../services/ClusterTagsFilter'
 
-import { useFetchPaginatedData } from '../utils/useFetchPaginatedData'
+import { useFetchPaginatedData } from '../../utils/table/useFetchPaginatedData'
 
 import { useProjectId } from '../../contexts/ProjectsContext'
