+++ assets/src/components/cd/globalServices/details/GlobalService.tsx
@@ -48,7 +48,7 @@ import { LinkTabWrap } from '../../../utils/Tabs'
 import { PluralErrorBoundary } from '../../PluralErrorBoundary'
 import KickButton from '../../../utils/KickButton'
 import { GqlError } from '../../../utils/Alert'
-import { useFetchPaginatedData } from '../../utils/useFetchPaginatedData'
+import { useFetchPaginatedData } from '../../../utils/table/useFetchPaginatedData'
 import { useProjectId } from '../../../contexts/ProjectsContext'
 import { mapExistingNodes } from '../../../../utils/graphql'
 import { DistroProviderIcon } from '../../../utils/ClusterDistro'
