+++ assets/src/components/cd/namespaces/NamespacesTable.tsx
@@ -12,7 +12,7 @@ import { GqlError } from 'components/utils/Alert'
 
 import { useProjectId } from 'components/contexts/ProjectsContext'
 
-import { useFetchPaginatedData } from '../utils/useFetchPaginatedData'
+import { useFetchPaginatedData } from '../../utils/table/useFetchPaginatedData'
 
 import { NAMESPACES_REACT_VIRTUAL_OPTIONS, columns } from './Namespaces'
