+++ assets/src/components/home/clusteroverview/ClusterOverviewCard.tsx
@@ -2,7 +2,7 @@ import { H1 } from 'honorable'
 
 import { useTheme } from 'styled-components'
 import { CLUSTERS_REACT_VIRTUAL_OPTIONS } from 'components/cd/clusters/Clusters'
-import { useFetchPaginatedData } from 'components/cd/utils/useFetchPaginatedData'
+import { useFetchPaginatedData } from 'components/utils/table/useFetchPaginatedData'
 import { GqlError } from 'components/utils/Alert'
 import { useClustersQuery, useUpgradeStatisticsQuery } from 'generated/graphql'
