+++ assets/src/components/home/deployments/DeploymentsCard.tsx
@@ -1,4 +1,4 @@
-import { useFetchPaginatedData } from 'components/cd/utils/useFetchPaginatedData'
+import { useFetchPaginatedData } from 'components/utils/table/useFetchPaginatedData'
 
 import {
   ServiceDeploymentStatus,
