+++ assets/src/components/pr/scm/PrScmConnections.tsx
@@ -10,7 +10,7 @@ import { useSetPageHeaderContent } from 'components/cd/ContinuousDeployment'
 
 import { PR_BASE_CRUMBS, PR_SCM_ABS_PATH } from 'routes/prRoutesConsts'
 
-import { useFetchPaginatedData } from 'components/cd/utils/useFetchPaginatedData'
+import { useFetchPaginatedData } from 'components/utils/table/useFetchPaginatedData'
 
 import { columns } from './PrScmConnectionsColumns'
 import { CreateScmConnection } from './CreateScmConnection'
