+++ assets/src/components/settings/usermanagement/groups/GroupsList.tsx
@@ -4,7 +4,7 @@ import { useGroupsQuery } from 'generated/graphql'
 import isEmpty from 'lodash/isEmpty'
 import { ComponentProps, useContext, useMemo } from 'react'
 
-import { useFetchPaginatedData } from 'components/cd/utils/useFetchPaginatedData'
+import { useFetchPaginatedData } from 'components/utils/table/useFetchPaginatedData'
 
 import { GqlError } from 'components/utils/Alert'
