+++ assets/src/components/settings/usermanagement/serviceaccounts/ServiceAccountsList.tsx
@@ -6,7 +6,7 @@ import { useServiceAccountsQuery } from 'generated/graphql'
 
 import LoadingIndicator from 'components/utils/LoadingIndicator'
 
-import { useFetchPaginatedData } from 'components/cd/utils/useFetchPaginatedData'
+import { useFetchPaginatedData } from 'components/utils/table/useFetchPaginatedData'
 
 import { GqlError } from 'components/utils/Alert'
