+++ assets/src/components/utils/table/ColWithIcon.tsx
@@ -1,6 +1,6 @@
-import { AppIcon, Tooltip, WrapWithIf } from '@pluralsh/design-system'
+import { AppIcon, WrapWithIf } from '@pluralsh/design-system'
 import styled from 'styled-components'
-import { ComponentProps, ReactNode } from 'react'
+import { ComponentProps } from 'react'
 import classNames from 'classnames'
 import { Merge } from 'type-fest'
 
@@ -26,55 +26,6 @@ const ColWithIconSC = styled.div(({ theme }) => ({
     },
   },
 }))
-
-export function ColWithOptionalIcon({
-  icon,
-  iconTooltip,
-  children,
-  truncateLeft = false,
-  ...props
-}: Merge<
-  ComponentProps<typeof ColWithIconSC>,
-  {
-    icon?: string | ComponentProps<typeof AppIcon>['icon']
-    truncateLeft?: boolean
-    iconTooltip?: ReactNode
-  }
->) {
-  let iconElt = !!icon && (
-    <AppIcon
-      spacing="padding"
-      size="xxsmall"
-      icon={typeof icon !== 'string' ? icon : undefined}
-      url={typeof icon === 'string' ? icon : undefined}
-    />
-  )
-
-  if (icon && iconTooltip) {
-    iconElt = (
-      <Tooltip
-        placement="top"
-        label={iconTooltip}
-      >
-        <span>{iconElt}</span>
-      </Tooltip>
-    )
-  }
-
-  return (
-    <ColWithIconSC {...props}>
-      {iconElt && <div className="icon">{iconElt}</div>}
-      <div className={classNames('content', { truncateLeft: 'truncateLeft' })}>
-        <WrapWithIf
-          condition={truncateLeft}
-          wrapper={<span />}
-        >
-          {children}
-        </WrapWithIf>
-      </div>
-    </ColWithIconSC>
-  )
-}
 
 export function ColWithIcon({
   icon,
